[sdk, simplex]: move phantom aggregation back into the SDK; add multi-pair filler E2E test
- Move aggregatePhantomBids orchestration from the indexer back into the SDK so it isn't
  duplicated (it was reimplemented in the indexer and mirrored in the simplex E2E test).
  Bid decoding is injectable via the `extractFill` param: the indexer passes its VM2-safe
  ethers decoder (extractFillDataVm2), while the default stays the viem-based extractFillData
  for Node consumers (tests, simplex). getTotalSolverBalance/sweepSolverLiquidity are
  internal to the SDK again; the indexer handler and the E2E test call aggregatePhantomBids
  directly.
- Add an E2E regression test for the per-pair retraction fix: configure two pairs on one
  chain (USDC->cNGN and cNGN->USDC) and assert every phantom order keeps a full set of bids,
  i.e. bidding on the second pair no longer retracts the first pair's bid.
